requirements.txt
fastapi
uvicorn[standard]
aiomysql
pydantic
python-dotenv
